+++ workers/unity/Packages/io.improbable.gdk.buildsystem/Configuration/BuildContext.cs
@@ -12,10 +12,11 @@ namespace Improbable.Gdk.BuildSystem.Configuration
         public BuildTargetConfig BuildTargetConfig;
         public BuildEnvironment BuildEnvironment;
         public ScriptingImplementation ScriptingImplementation;
+        public iOSSdkVersion? IOSSdkVersion;
 
         public static List<BuildContext> GetBuildContexts(IEnumerable<string> wantedWorkerTypes,
             BuildEnvironment buildEnvironment, ScriptingImplementation? scriptImplementation = null,
-            ICollection<BuildTarget> buildTargetFilter = null)
+            ICollection<BuildTarget> buildTargetFilter = null, iOSSdkVersion? iosSdkVersion = null)
         {
             var spatialOsBuildConfiguration = BuildConfig.GetInstance();
             var result = new List<BuildContext>();
@@ -60,7 +61,8 @@ namespace Improbable.Gdk.BuildSystem.Configuration
                     WorkerType = workerType, BuildEnvironment = buildEnvironment,
                     ScriptingImplementation = scriptImplementation ??
                         PlayerSettings.GetScriptingBackend(BuildPipeline.GetBuildTargetGroup(targetConfig.Target)),
-                    BuildTargetConfig = targetConfig
+                    BuildTargetConfig = targetConfig,
+                    IOSSdkVersion = (targetConfig.Target == BuildTarget.iOS) ? iosSdkVersion : null
                 }));
             }